Move task_list components into directory
