+++ static/statuses.json
@@ -43,9 +43,9 @@
   "Primer::SubheadComponent": "beta",
   "Primer::TabContainerComponent": "alpha",
   "Primer::TabNavComponent": "alpha",
-  "Primer::TaskList": "alpha",
-  "Primer::TaskListItem": "alpha",
-  "Primer::TaskListList": "alpha",
+  "Primer::TaskList::Item": "alpha",
+  "Primer::TaskList::List": "alpha",
+  "Primer::TaskListComponent": "alpha",
   "Primer::TextComponent": "beta",
   "Primer::TimeAgoComponent": "beta",
   "Primer::TimelineItemComponent": "beta",
