+++ test/components/component_test.rb
@@ -7,9 +7,9 @@ class PrimerComponentTest < Minitest::Test
 
   # Components with any arguments necessary to make them render
   COMPONENTS_WITH_ARGS = [
-    [Primer::TaskList, {}],
-    [Primer::TaskListList, {}],
-    [Primer::TaskListItem, {}],
+    [Primer::TaskListComponent, {}],
+    [Primer::TaskList::List, {}],
+    [Primer::TaskList::Item, {}],
     [Primer::LocalTime, { datetime: DateTime.parse("2014-06-01T13:05:07Z") }],
     [Primer::ImageCrop, { src: "Foo" }],
     [Primer::IconButton, { icon: :star, "aria-label": "Label" }],
